add change not foound exec

+++ src/main/java/com/app/testingService/repos/PersonRepo.java
@@ -8,8 +8,9 @@ import com.app.testingService.models.Person;
 import reactor.core.publisher.Mono;
 
 @Repository
-public interface PersonRepo extends ReactiveCrudRepository<Person,Integer> {
+public interface PersonRepo extends ReactiveCrudRepository<Person,Long> {
     
     Mono<Person> findByUsername(String username);
     Mono<Person> findById(Long id);
+    
 }
